Switched writer and reader back to int (hw3 - lzw) (wip)

+++ Zipper/Zipper.Tests/ArbitraryBitReaderWriterTests.cs
@@ -44,7 +44,7 @@ public class ArbitraryBitReaderWriterTests
             {
                 Assert.Multiple(() =>
                 {
-                    Assert.That(reader.ReadNext(out uint number), Is.True);
+                    Assert.That(reader.ReadNext(out int number), Is.True);
                     Assert.That(number, Is.EqualTo(data.Numbers[i]));
                 });
             }
@@ -64,7 +64,7 @@ public class ArbitraryBitReaderWriterTests
             using var writer = new ArbitraryBitWriter(memory, width);
             for (int i = 0; i < numbersCount; i++)
             {
-                writer.Write((uint)(i * i));
+                writer.Write(i * i);
             }
         }
 
@@ -75,7 +75,7 @@ public class ArbitraryBitReaderWriterTests
             {
                 Assert.Multiple(() =>
                 {
-                    Assert.That(reader.ReadNext(out uint number), Is.True);
+                    Assert.That(reader.ReadNext(out int number), Is.True);
                     Assert.That(number, Is.EqualTo(i * i));
                 });
             }
@@ -98,10 +98,10 @@ public class ArbitraryBitReaderWriterTests
             int width = i + minWidth;
             long upperBound = 1L << width;
 
-            var numbers = new uint[numbersLength];
+            var numbers = new int[numbersLength];
             for (int j = 0; j < numbersLength; j++)
             {
-                numbers[j] = (uint)random.NextInt64(upperBound);
+                numbers[j] = (int)random.NextInt64(upperBound);
             }
 
             result[i] = new(width, numbers);
@@ -110,7 +110,7 @@ public class ArbitraryBitReaderWriterTests
         return result;
     }
 
-    public readonly record struct TestData(int Width, uint[] Numbers);
+    public readonly record struct TestData(int Width, int[] Numbers);
 
     // use MemoryStream, because implementing all Stream's abstract membrers leads to bad code coverage
     private class TestStream : MemoryStream
